Mocha testing setup for the firebase dir, emulators set up and working with mocha
