+++ firebase/functions/database-debug.log
@@ -3,7 +3,11 @@ WARNING: Illegal reflective access by io.netty.util.internal.ReflectionUtil (fil
 WARNING: Please consider reporting this to the maintainers of io.netty.util.internal.ReflectionUtil
 WARNING: Use --illegal-access=warn to enable warnings of further illegal reflective access operations
 WARNING: All illegal access operations will be denied in a future release
-19:52:10.576 [NamespaceSystem-akka.actor.default-dispatcher-4] INFO akka.event.slf4j.Slf4jLogger - Slf4jLogger started
-19:52:10.862 [main] INFO com.firebase.server.forge.App$ - Listening at 127.0.0.1:9000
-19:52:24.790 [NamespaceSystem-akka.actor.default-dispatcher-4] INFO com.firebase.core.namespace.NamespaceActor - organizer-9cce6-default-rtdb successfully activated FBKV (SurveyIdle(0)) wait: 287ms, init: 0ms
-19:52:24.862 [NamespaceSystem-blocking-namespace-operation-dispatcher-6] INFO com.firebase.core.namespace.StateManager - Namespace organizer-9cce6-default-rtdb status Active to Active
+21:34:25.141 [NamespaceSystem-akka.actor.default-dispatcher-4] INFO akka.event.slf4j.Slf4jLogger - Slf4jLogger started
+21:34:25.376 [main] INFO com.firebase.server.forge.App$ - Listening at 127.0.0.1:9000
+21:34:37.317 [NamespaceSystem-akka.actor.default-dispatcher-4] INFO com.firebase.core.namespace.NamespaceActor - organizer-default-rtdb successfully activated FBKV (SurveyIdle(0)) wait: 215ms, init: 0ms
+21:34:37.449 [NamespaceSystem-blocking-namespace-operation-dispatcher-6] INFO com.firebase.core.namespace.StateManager - Namespace organizer-default-rtdb status Active to Active
+21:34:44.396 [Thread-0] INFO com.firebase.server.forge.App$ - Attempting graceful shutdown.
+21:34:44.405 [NamespaceSystem-akka.actor.default-dispatcher-5] INFO com.firebase.core.namespace.Terminator$Terminator - 1 actors left to terminate: organizer-default-rtdb
+21:34:44.417 [NamespaceSystem-akka.actor.default-dispatcher-4] INFO com.firebase.core.namespace.NamespaceActor - stopped namespace actor for organizer-default-rtdb
+21:34:44.429 [Thread-0] INFO com.firebase.server.forge.App$ - Graceful shutdown complete.
