+++ firebase/functions/firestore-debug.log
@@ -1,4 +1,4 @@
-Jul 24, 2023 7:52:07 PM com.google.cloud.datastore.emulator.firestore.websocket.WebSocketServer start
+Jul 24, 2023 9:34:22 PM com.google.cloud.datastore.emulator.firestore.websocket.WebSocketServer start
 INFO: Started WebSocket server on ws://127.0.0.1:9150
 API endpoint: http://127.0.0.1:8080
 If you are using a library that supports the FIRESTORE_EMULATOR_HOST environment variable, run:
@@ -7,5 +7,5 @@ If you are using a library that supports the FIRESTORE_EMULATOR_HOST environment
 
 Dev App Server is now running.
 
-Jul 24, 2023 7:53:06 PM com.google.cloud.datastore.emulator.firestore.websocket.WebSocketChannelHandler initChannel
-INFO: Connected to new websocket client
+*** shutting down gRPC server since JVM is shutting down
+*** server shut down
