+++ firebase/functions/package.json
@@ -8,7 +8,9 @@
     "shell": "npm run build && firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
+    "logs": "firebase functions:log",
+    "test": "mocha --require ts-node/register src/tests/**/*.spec.ts --recursive",
+    "test:emulators": "firebase emulators:exec --ui --project=organizer 'npm run test'"
   },
   "engines": {
     "node": "18"
@@ -19,12 +21,16 @@
     "firebase-functions": "^4.3.1"
   },
   "devDependencies": {
+    "@types/mocha": "^10.0.1",
     "@typescript-eslint/eslint-plugin": "^5.12.0",
     "@typescript-eslint/parser": "^5.12.0",
+    "chai": "^4.3.7",
     "eslint": "^8.9.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-import": "^2.25.4",
     "firebase-functions-test": "^3.1.0",
+    "mocha": "^10.2.0",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.0"
   },
   "private": true
